requirements.txt
Flask==3.0.3
flask-cors==4.0.0
opencv-python==4.10.0.84
torch==2.2.2
torchvision==0.17.2
pytesseract==0.3.13
numpy==1.26.4
Pillow==10.2.0
Unidecode==1.3.8
python-dotenv==1.0.1
pyzbar==0.1.9